requirements_retrieval.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.14
albumentations==1.4.14
all-clip==1.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
autofaiss==2.17.0
blinker==1.8.2
braceexpand==0.1.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip-anytorch==2.6.0
clip-retrieval==2.44.0
confection==0.1.5
dataclasses==0.6
datasets==1.18.4
docker-pycreds==0.4.0
embedding-reader==1.7.0
eval_type_backport==0.2.0
ExifRead-nocycle==3.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
fire==0.5.0
Flask==3.0.3
Flask-Cors==4.0.2
Flask-RESTful==0.3.10
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
img2dataset==1.45.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
multidict==6.0.5
multilingual-clip==1.0.10
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.13.2
segment-anything-hq==0.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
tensorboardX==2.6.2.2
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.9
toml==0.10.2
#torch==2.1.2+cu121
#torchaudio==2.1.2+cu121
#torchvision==0.16.2+cu121
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.0
triton==2.1.0
ttach==0.0.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
wandb==0.16.6
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
Werkzeug==3.0.4
wrapt==1.16.0
xformers==0.0.23.post1
xxhash==3.5.0
yarl==1.9.11
zipp==3.20.1
dreamsim